# CBG-1084: Optimize RevTree Unmarshal