- Preallocate `RevTree` map size
- Reorder `RevInfo` struct fields for memory alignment

```benchmark                                            old ns/op     new ns/op     delta
BenchmarkRevtreeUnmarshal/Marshal_into_revTree-8     4453516       4468108       +0.33%

benchmark                                            old allocs     new allocs     delta
BenchmarkRevtreeUnmarshal/Marshal_into_revTree-8     16791          16718          -0.43%

benchmark                                            old bytes     new bytes     delta
BenchmarkRevtreeUnmarshal/Marshal_into_revTree-8     1398797       1269252       -9.26%
```